v2.0.1 -> v2.0.2, appeased rust 1.66

+++ src/ui/components/popups/mod.rs
@@ -13,7 +13,7 @@ pub fn centered_popup(size: Rect, terminal_height: u16) -> Rect {
         .direction(Direction::Vertical)
         .constraints(
             [
-                Constraint::Length((terminal_height / 2) as u16 - 6),
+                Constraint::Length((terminal_height / 2) - 6),
                 Constraint::Length(3),
                 Constraint::Min(0),
             ]
